ExtractText: handle text matrix (Tm) boxes
Previously Tm boxes would have their text extracted, but with no
formatting information between boxes. Now, if a box is lower on the
page than has previously been recorded, we insert a newline, and if
a box is further to the right on the same line than a previous box, we
insert a tab character.

This works well enough for parsing Clipper Card transaction histories.
You can see an example at https://clipper-csv.appspot.com.

Fixes #141.